Function rustchat::netlib::read

pub fn read(client: &TcpStream, data: &mut String) -> usize

read data from client

argument

client - reference to a TcpStream

data - reference to a mutable String

return value

return the number of bytes received from the connection

Example :

// setup the variable containing the address to connect to
let conn_id = format!("{}:{}", "127.0.0.1".to_string(), "6666".to_string());
// start the tcp connection
let conn = listen(&conn_id);
// accept a client
let client = accept(listener);
let mut client_data = String::new();
read(&client, &mut client_data);